requirements.txt
python-dotenv
requests
python-telegram-bot
flake8
Pillow
django==3.2